Refactor smuggler import to DatabaseDataImporter
